requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.34
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
jinja2==3.1.4
httpx==0.27.2
phonenumberslite==8.13.47